LPD-104914 Move DSRRoleAssignmentPermissionUtil to the util package
